Move utility function to separate file and module
